pyproject.toml
[project]
name = "grawlix"
authors = [
    { name = "Joakim Holm", email = "mail@joakimholm.xyz" }
]
description = "CLI tool for downloading ebooks"
readme = "README.md"
keywords = ["ebook", "cli", "downloader"]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]
dependencies = [
    "beautifulsoup4>=4.9.0",
    "bbpb>=1.0.0",
    "EbookLib>=0.17",
    "httpx>=0.23.0",
    "importlib-resources>=5.0",
    "lxml>=4.6.0",
    "platformdirs>=3.0.0",
    "pycryptodome>=3.10.0",
    "rich>=10.0.0",
    "tomli>=1.0.0; python_version<'3.11'",
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-httpx>=0.21.0",
]

[project.urls]
"Homepage" = "https://github.com/jo1gi/grawlix"
"Bugtracker" = "https://github.com/jo1gi/grawlix/issues"

[project.scripts]
grawlix = "grawlix.__main__:run"

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"


[tool.setuptools.dynamic]
version = {attr = "grawlix.__version__"}

[tool.setuptools.package-data]
grawlix = ["*.txt"]


[tool.mypy]
ignore_missing_imports = true
allow_untyped_globals = false
disallow_untyped_calls = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--cov=grawlix",
    "--cov-report=html",
    "--cov-report=term-missing:skip-covered",
    "-v",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
